Refactors Dockerfile for smaller image size
Refactors the Dockerfile to use a multi-stage build, resulting in a smaller final image size.

The first stage uses a uv base image to install dependencies and the second stage uses an Alpine Linux base image to run the application.
This reduces the final image size and improves security by minimizing the attack surface.